fix: share Goose state path resolution
